Safari-609.1.11.1/LayoutTests/fast/css/border-width-large-expected.txt

Test for WebKit bug 18294 : Strange Result for getComputedStyle on borderWidth set in em

PASS computedStyle.getPropertyValue('border-width') is '12000px 11000px 10000px 9010px'
PASS computedStyle.getPropertyCSSValue('border-top-width').cssText is '12000px'
PASS computedStyle.getPropertyCSSValue('border-right-width').cssText is '11000px'
PASS computedStyle.getPropertyCSSValue('border-bottom-width').cssText is '10000px'
PASS computedStyle.getPropertyCSSValue('border-left-width').cssText is '9010px'
PASS successfullyParsed is true

TEST COMPLETE
